added exception handling

+++ src/main/java/com/aquent/pizzaorder/dao/impl/PizzaOrderDaoImpl.java
@@ -13,11 +13,13 @@ import java.util.TreeSet;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Repository;
 
 import com.aquent.pizzaorder.dao.IPizzaOrderDao;
+import com.aquent.pizzaorder.exceptions.ApplicationException;
 import com.aquent.pizzaorder.model.PizzaOrderDto;
 
 import lombok.extern.slf4j.Slf4j;
@@ -25,10 +27,7 @@ import lombok.extern.slf4j.Slf4j;
 @Repository
 @Slf4j
 public class PizzaOrderDaoImpl implements IPizzaOrderDao {
-	/*@Value("${systemProperties['source.path']?:'sampl'}")*/
 	private String soursePath;
-	
-	/*@Value("${systemProperties['destination.path']?:'sample'}")*/
 	private String destinationPath;
 	
 	@Autowired
@@ -41,9 +40,15 @@ public class PizzaOrderDaoImpl implements IPizzaOrderDao {
 	public void loadPizzaOrdersData() {
 		log.info("Destination Path ::::::::::::::::::: "+soursePath);
 		log.info("Source Path ::::::::::::::::::: "+destinationPath);
+		
+		if(StringUtils.isEmpty(soursePath) || StringUtils.isEmpty(destinationPath)) {
+			throw new ApplicationException("PATH_ERROR", "Source Path or Destination Path Not Valid");
+		}
+		
 		File file = new File(destinationPath);
 		if(file.exists())
 			file.delete();
+		
 		try (	
 				FileWriter fw = new FileWriter(file, true);
 				BufferedWriter bw = new BufferedWriter(fw);
@@ -62,7 +67,7 @@ public class PizzaOrderDaoImpl implements IPizzaOrderDao {
 			    out.flush();
 			});
 		} catch (IOException e) {
-			e.printStackTrace();
+			throw new ApplicationException("FILE_OPERATIONS", e.getMessage());
 		} 
 	}
 	
@@ -78,7 +83,7 @@ public class PizzaOrderDaoImpl implements IPizzaOrderDao {
 				  return new PizzaOrderDto(orderDetails[0], Long.valueOf(orderDetails[1]));
 			  }).collect(Collectors.toCollection(TreeSet::new));
 		} catch (IOException e) {
-			e.printStackTrace();
+			throw new ApplicationException("FILE_OPERATIONS", e.getMessage());
 		} 
 		return pizzaOrders;
 	}
@@ -97,7 +102,7 @@ public class PizzaOrderDaoImpl implements IPizzaOrderDao {
 		    out.println(pizzaOrderDto.getName()+","+pizzaOrderDto.getTime());
 		    out.flush();
 		} catch (IOException e) {
-			e.printStackTrace();
+			throw new ApplicationException("FILE_OPERATIONS", e.getMessage());
 		} 
 		return pizzaOrderDto;
 	}
